requirements.txt
flask==2.0.1
openai==0.27.0
python-dotenv==0.19.0
requests==2.26.0
google-api-python-client==2.45.0
gunicorn==20.1.0
